Add route for login
Had to add a slightly hacky function to the already slightly hacky promise wrapper.If my purpose it to be able to reuse it later on with other systems, then I might need to start again. Or just build something similar for said other systems.

Two hacks make a feature?