Remove dev container
Well, it was a good idea at first, but a couple of quirks just ekpt getting on my nerves. I'll do some more playing around later, but for now I'll just remove it. There doesn't seem to be a setting to just disable it, and commenting out most of the file causes errors to be thrown that are also a pain, so deletion was the easiest thing to do