Remove unused logger from IdentiferService
Also fixes issue where an unusual mock is required in some of the tests.

This was actually a pretty difficult one to pin down. These services were all importing, through different chains, the ApiError. The error class uses the IdentifierService to generate an ID. Since it gets the service by using Container.get(), it wasn't able to be mocked properly. This led to the actual real logger being used as well.

Removing the (unused) logger from the IdentifierService does not really fix the problem, only making the side effects less impactful.

There are going to be some related issues when trying to test the express middleware